+++ LearnTorch/Core/__pycache__/Square.cpython-39.pyc
