+++ LearnTorch/Test/TestBackward.py
@@ -1,6 +1,6 @@
 #用于测试各函数的文件
 import numpy as np
-from Core.Variable import Variable
+from Core.VariableFunction import Variable
 from Core.Square import Square
 from Core.Exp import Exp
