+++ LearnTorch/Test/TestGeneration.py
@@ -1,9 +1,8 @@
 import unittest
 import numpy as np
-from Core.Variable import Variable
-from Core.Function import Function
+from Core.VariableFunction import Variable,Function,add
 from Core.Square import square
-from Core.Add import add
+
 
 #测试类，测试复杂计算图的正向传播和反向传播是否正确
 class SquareTest(unittest.TestCase):
